#2678 - Set default annotation editor in project settings
- Move project initializers to separate module to improve dependency graph
