+++ inception/inception-project-initializers/pom.xml
@@ -15,15 +15,17 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>de.tudarmstadt.ukp.inception.app</groupId>
     <artifactId>inception-app</artifactId>
     <version>22.0-SNAPSHOT</version>
   </parent>
-  <artifactId>inception-project</artifactId>
-  <name>INCEpTION - Core - Project</name>
+  <artifactId>inception-project-initializers</artifactId>
+  <name>INCEpTION - Core - Project initializers</name>
   <dependencies>
     <dependency>
       <groupId>commons-io</groupId>
@@ -40,11 +42,19 @@
 
     <dependency>
       <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+      <artifactId>inception-project</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>de.tudarmstadt.ukp.inception.app</groupId>
       <artifactId>inception-model</artifactId>
     </dependency>
     <dependency>
       <groupId>de.tudarmstadt.ukp.inception.app</groupId>
       <artifactId>inception-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+      <artifactId>inception-api-annotation</artifactId>
     </dependency>
     <dependency>
       <groupId>de.tudarmstadt.ukp.inception.app</groupId>
@@ -126,54 +136,5 @@
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
     </dependency>
-    
-    <!-- DEPENDENCIES FOR TESTING -->
-    <dependency>
-    	<groupId>org.springframework.boot</groupId>
-    	<artifactId>spring-boot</artifactId>
-    	<scope>test</scope>
-    </dependency>
-    <dependency>
-    	<groupId>org.springframework.boot</groupId>
-    	<artifactId>spring-boot-test-autoconfigure</artifactId>
-    	<scope>test</scope>
-    </dependency>
-    <dependency>
-    	<groupId>org.springframework.boot</groupId>
-    	<artifactId>spring-boot-starter-data-jpa</artifactId>
-    	<scope>test</scope>
-    </dependency>
-    <dependency>
-    	<groupId>com.h2database</groupId>
-    	<artifactId>h2</artifactId>
-    	<scope>test</scope>
-    </dependency>
   </dependencies>
-  <build>
-    <plugins>
-	  <plugin>
-	    <groupId>org.apache.maven.plugins</groupId>
-		<artifactId>maven-dependency-plugin</artifactId>
-		<executions>
-		  <execution>
-		    <id>default</id>
-			<phase>verify</phase>
-			<goals>
-			  <goal>analyze-only</goal>
-			</goals>
-		  </execution>
-		</executions>
-		<configuration>
-		  <failOnWarning>true</failOnWarning>
-		  <ignoredDependencies>
-			<!-- - h2database and data-jpa, used via reflection in tests, cannot 
-			     -be detected by Maven 
-			-->
-			<dependency>com.h2database:h2</dependency>
-			<dependency>org.springframework.boot:spring-boot-starter-data-jpa</dependency>
-		  </ignoredDependencies>
-		</configuration>
-	  </plugin>
-	</plugins>
-  </build>
 </project>
