+++ inception/inception-project-initializers/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/initializers/CoreferenceRelationTagSetInitializer.java
@@ -17,7 +17,7 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.project.initializers;
 
-import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.JsonImportUtil.importTagSetFromJson;
+import static de.tudarmstadt.ukp.clarin.webanno.project.initializers.JsonImportUtil.importTagSetFromJson;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -34,19 +34,19 @@ import de.tudarmstadt.ukp.clarin.webanno.project.initializers.config.ProjectInit
 /**
  * <p>
  * This class is exposed as a Spring Component via
- * {@link ProjectInitializersAutoConfiguration#namedEntityTagSetInitializer}.
+ * {@link ProjectInitializersAutoConfiguration#coreferenceRelationTagSetInitializer}.
  * </p>
  */
-public class NamedEntityTagSetInitializer
+public class CoreferenceRelationTagSetInitializer
     implements TagSetInitializer
 {
     // Name must match the name in the tag set file loaded by this initializer!
-    public static final String TAG_SET_NAME = "Named Entity tags";
+    public static final String TAG_SET_NAME = "Coreference relations";
 
     private final AnnotationSchemaService annotationSchemaService;
 
     @Autowired
-    public NamedEntityTagSetInitializer(AnnotationSchemaService aAnnotationSchemaService)
+    public CoreferenceRelationTagSetInitializer(AnnotationSchemaService aAnnotationSchemaService)
     {
         annotationSchemaService = aAnnotationSchemaService;
     }
@@ -73,7 +73,7 @@ public class NamedEntityTagSetInitializer
     public void configure(Project aProject) throws IOException
     {
         importTagSetFromJson(aProject,
-                new ClassPathResource("/tagsets/de-ne-webanno.json").getInputStream(),
+                new ClassPathResource("/tagsets/de-coref-rel-tuebadz.json").getInputStream(),
                 annotationSchemaService);
     }
 }
