+++ inception/inception-project-initializers/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/initializers/DependencyTypeTagSetInitializer.java
@@ -17,7 +17,7 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.project.initializers;
 
-import static de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.JsonImportUtil.importTagSetFromJson;
+import static de.tudarmstadt.ukp.clarin.webanno.project.initializers.JsonImportUtil.importTagSetFromJson;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -34,19 +34,19 @@ import de.tudarmstadt.ukp.clarin.webanno.project.initializers.config.ProjectInit
 /**
  * <p>
  * This class is exposed as a Spring Component via
- * {@link ProjectInitializersAutoConfiguration#coreferenceTypeTagSetInitializer}.
+ * {@link ProjectInitializersAutoConfiguration#dependencyTypeTagSetInitializer}.
  * </p>
  */
-public class CoreferenceTypeTagSetInitializer
+public class DependencyTypeTagSetInitializer
     implements TagSetInitializer
 {
     // Name must match the name in the tag set file loaded by this initializer!
-    public static final String TAG_SET_NAME = "Coreference mentions";
+    public static final String TAG_SET_NAME = "UD Universal Dependencies (v2)";
 
     private final AnnotationSchemaService annotationSchemaService;
 
     @Autowired
-    public CoreferenceTypeTagSetInitializer(AnnotationSchemaService aAnnotationSchemaService)
+    public DependencyTypeTagSetInitializer(AnnotationSchemaService aAnnotationSchemaService)
     {
         annotationSchemaService = aAnnotationSchemaService;
     }
@@ -73,7 +73,7 @@ public class CoreferenceTypeTagSetInitializer
     public void configure(Project aProject) throws IOException
     {
         importTagSetFromJson(aProject,
-                new ClassPathResource("/tagsets/de-coref-type-bart.json").getInputStream(),
+                new ClassPathResource("/tagsets/mul-dep-ud2.json").getInputStream(),
                 annotationSchemaService);
     }
 }
