+++ inception/inception-project-initializers/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/initializers/JsonImportUtil.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.api.annotation.util;
+package de.tudarmstadt.ukp.clarin.webanno.project.initializers;
 
 import java.io.IOException;
 import java.io.InputStream;
